+++ src/pages/Control/components/FloorFilter/components/Selected.tsx
@@ -1,0 +1,49 @@
+import React, { CSSProperties, useMemo } from 'react'
+
+import s from './Selected.module.css'
+
+type SelectDirect = 'LEFT' | 'RIGHT'
+export default function Selected({
+  dragStartPoint,
+  dragEndPoint,
+  intervalWidth,
+  backgroundColor,
+}: {
+  dragStartPoint: number
+  dragEndPoint: number
+  intervalWidth: number
+  backgroundColor?: CSSProperties['backgroundColor']
+}) {
+  const direct = useMemo<SelectDirect>(() => {
+    if (dragEndPoint > dragStartPoint) {
+      return 'RIGHT'
+    } else {
+      return 'LEFT'
+    }
+  }, [dragEndPoint, dragStartPoint])
+
+  const selectedWidth = useMemo(() => {
+    const len = Math.abs(dragEndPoint - dragStartPoint)
+    return len * intervalWidth
+  }, [dragEndPoint, dragStartPoint, intervalWidth])
+
+  const selectedLeft = useMemo(() => {
+    if (direct === 'LEFT') {
+      return dragEndPoint * intervalWidth
+    } else {
+      return dragStartPoint * intervalWidth
+    }
+  }, [direct, dragEndPoint, dragStartPoint, intervalWidth])
+
+  return (
+    <div className={s.SelectedContainer} style={{ backgroundColor }}>
+      <div
+        className={s.Selected}
+        style={{
+          width: `${selectedWidth}px`,
+          left: `${selectedLeft}px`
+        }}
+      ></div>
+    </div>
+  )
+}
