+++ src/pages/Control/components/FloorFilter/hooks/useMouseDrag.ts
@@ -1,0 +1,71 @@
+import { useEffect, useState } from 'react'
+
+export default function useMouseDrag({
+  interval_width,
+  onClick,
+  onDragEnd,
+}: {
+  interval_width: number
+  onClick: (p: number) => void
+  onDragEnd: (s: number, e: number) => void
+}) {
+  const [drag_start_point, setDragStartPoint] = useState(0)
+  const [drag_end_point, setDragEndPoint] = useState(0)
+
+  const [mouse_move_start, setMouseMoveStart] = useState<number | null>(null)
+  const [mouse_move_offset, setMouseMoveOffset] = useState<number | null>(null)
+
+  useEffect(() => {
+    if (mouse_move_start !== null) {
+      const handler = (e: MouseEvent) => setMouseMoveOffset( e.clientX - mouse_move_start )
+      window.addEventListener('mousemove', handler)
+      return () => window.removeEventListener('mousemove', handler)
+    }
+  }, [mouse_move_start])
+
+  useEffect(() => {
+    if (mouse_move_offset !== null) {
+      console.log('mouse_move_offset', mouse_move_offset)
+
+      const abs_mouse_move_offset = Math.abs(mouse_move_offset)
+
+      const point_offset = Math.floor(abs_mouse_move_offset / interval_width)
+      if (mouse_move_offset > 0) {
+        setDragEndPoint(drag_start_point + point_offset)
+      } else {
+        setDragEndPoint(drag_start_point - point_offset)
+      }
+    }
+  }, [drag_start_point, interval_width, mouse_move_offset])
+
+  useEffect(() => {
+    if (mouse_move_start !== null) {
+      const handler = (e: MouseEvent) => {
+        if (mouse_move_offset === null) {
+          // is click
+          onClick(drag_start_point)
+        } else {
+          // submit selected
+          console.log('dragend range', drag_start_point, drag_end_point)
+          onDragEnd(drag_start_point, drag_end_point)
+        }
+
+        // setDragStartPoint(0)
+        setDragEndPoint(drag_start_point)
+        setMouseMoveStart(null)
+        setMouseMoveOffset(null)
+      }
+      window.addEventListener('mouseup', handler)
+      return () => window.removeEventListener('mouseup', handler)
+    }
+  }, [drag_end_point, drag_start_point, mouse_move_offset, mouse_move_start, onClick, onDragEnd])
+
+  return {
+    drag_start_point, setDragStartPoint,
+    drag_end_point, setDragEndPoint,
+    mouse_move_start, setMouseMoveStart,
+    mouse_move_offset, setMouseMoveOffset,
+
+    is_dragging: mouse_move_start !== null
+  }
+}
