+++ src/pages/Control/components/FloorFilter/hooks/useOffsetWidth.ts
@@ -1,0 +1,15 @@
+import { useEffect, useRef, useState } from 'react'
+
+export default function useOffsetWidth<RefType extends HTMLElement>() {
+  const ref = useRef<RefType>(null)
+  const [offset_width, setOffsetWidth] = useState(0)
+
+  useEffect(() => {
+    const el = ref.current
+    if (el) {
+      setOffsetWidth(el.offsetWidth)
+    }
+  }, [])
+
+  return [offset_width, ref] as const
+}
