+++ src/pages/Control/components/FloorFilter/index.module.css
@@ -29,94 +29,3 @@
   height: calc(var(--track-height) + var(--text-height) + var(--point-size));
   user-select: none;
 }
-
-.SelectedContainer {
-  width: 100%;
-  background-color: var(--unselected-color);
-  height: var(--track-height);
-  position: absolute;
-  top: var(--TrackLayoutTop);
-  left: 0;
-}
-.Selected {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 100%;
-  width: 0px;
-  background-color: var(--selected-color);
-}
-
-.FloorTextLayout {
-  width: 100%;
-  position: relative;
-}
-
-.FloorText.FloorTextHighlight {
-  color: rgba(0, 0, 0, 0.65);
-}
-.FloorText {
-  position: absolute;
-  top: var(--TextLayoutTop);
-
-  cursor: default;
-
-  width: var(--text-width);
-  text-align: center;
-
-  font-family: Arial, Helvetica, sans-serif;
-  font-weight: 600;
-  color: var(--unactive-color);
-  font-size: 14px;
-  line-height: 1em;
-}
-
-.FloorPointLayout {
-  width: 100%;
-  position: relative;
-}
-
-.FloorPoint.FloorPointHighlight:before {
-  box-shadow: inset 0 1px 1px var(--unactive-color);
-  background: var(--filter-background-color);
-}
-.FloorPoint.FloorPointHighlight {
-  background: hsl(240deg 8% 35%);
-  border: 0;
-  box-shadow: 0 4px 4px -2px rgb(0 0 0 / 50%);
-}
-
-.FloorFilter.isDragging .FloorPoint {
-  cursor:inherit;
-}
-
-.FloorPoint:before {
-  content: " ";
-  display: block;
-  width: calc(var(--point-size) - (3px * 2 + 1px));
-  height: calc(var(--point-size) - (3px * 2 + 1px));
-  border-radius: var(--point-size);
-}
-.FloorPoint {
-  position: absolute;
-  top: var(--PointLayoutTop);
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  text-align: center;
-
-  width: var(--point-size);
-  height: var(--point-size);
-  border-radius: var(--point-size);
-  box-sizing: border-box;
-
-  background: var(--filter-background-color);
-  box-shadow:
-    inset 0 0px 2px rgb(0 0 0 / 20%),
-    inset 0 2px 2px rgb(0 0 0 / 20%)
-  ;
-
-  cursor: pointer;
-}
