+++ src/pages/Control/components/FloorFilter/index.tsx
@@ -1,5 +1,10 @@
 import { compose, range, remove, sort, uniq } from 'ramda'
-import React, { CSSProperties, useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import React, { CSSProperties, useCallback, useEffect, useMemo, useState } from 'react'
+
+import PointAndText from './components/PointAndText'
+import Selected from './components/Selected'
+import useMouseDrag from './hooks/useMouseDrag'
+import useOffsetWidth from './hooks/useOffsetWidth'
 
 import s from './index.module.css'
 
@@ -41,184 +46,6 @@ function rangeToFloors([start, end]: Range) {
       return range(end, start + 1)
     }
   }
-}
-
-function useOffsetWidth<RefType extends HTMLElement>() {
-  const ref = useRef<RefType>(null)
-  const [offset_width, setOffsetWidth] = useState(0)
-
-  useEffect(() => {
-    const el = ref.current
-    if (el) {
-      setOffsetWidth(el.offsetWidth)
-    }
-  }, [])
-
-  return [offset_width, ref] as const
-}
-
-function useMouseDrag({
-  interval_width,
-  onClick,
-  onDragEnd,
-}: {
-  interval_width: number
-  onClick: (p: number) => void
-  onDragEnd: (s: number, e: number) => void
-}) {
-  const [drag_start_point, setDragStartPoint] = useState(0)
-  const [drag_end_point, setDragEndPoint] = useState(0)
-
-  const [mouse_move_start, setMouseMoveStart] = useState<number | null>(null)
-  const [mouse_move_offset, setMouseMoveOffset] = useState<number | null>(null)
-
-  useEffect(() => {
-    if (mouse_move_start !== null) {
-      const handler = (e: MouseEvent) => setMouseMoveOffset( e.clientX - mouse_move_start )
-      window.addEventListener('mousemove', handler)
-      return () => window.removeEventListener('mousemove', handler)
-    }
-  }, [mouse_move_start])
-
-  useEffect(() => {
-    if (mouse_move_offset !== null) {
-      console.log('mouse_move_offset', mouse_move_offset)
-
-      const abs_mouse_move_offset = Math.abs(mouse_move_offset)
-
-      const point_offset = Math.floor(abs_mouse_move_offset / interval_width)
-      if (mouse_move_offset > 0) {
-        setDragEndPoint(drag_start_point + point_offset)
-      } else {
-        setDragEndPoint(drag_start_point - point_offset)
-      }
-    }
-  }, [drag_start_point, interval_width, mouse_move_offset])
-
-  useEffect(() => {
-    if (mouse_move_start !== null) {
-      const handler = (e: MouseEvent) => {
-        if (mouse_move_offset === null) {
-          // is click
-          onClick(drag_start_point)
-        } else {
-          // submit selected
-          console.log('dragend range', drag_start_point, drag_end_point)
-          onDragEnd(drag_start_point, drag_end_point)
-        }
-
-        // setDragStartPoint(0)
-        setDragEndPoint(drag_start_point)
-        setMouseMoveStart(null)
-        setMouseMoveOffset(null)
-      }
-      window.addEventListener('mouseup', handler)
-      return () => window.removeEventListener('mouseup', handler)
-    }
-  }, [drag_end_point, drag_start_point, mouse_move_offset, mouse_move_start, onClick, onDragEnd])
-
-  return {
-    drag_start_point, setDragStartPoint,
-    drag_end_point, setDragEndPoint,
-    mouse_move_start, setMouseMoveStart,
-    mouse_move_offset, setMouseMoveOffset,
-
-    is_dragging: mouse_move_start !== null
-  }
-}
-
-function PointAndText({
-  floors,
-  highlightList,
-  intervalWidth,
-  onMouseDownPoint,
-}: {
-  floors: number[]
-  highlightList: number[]
-  intervalWidth: Exclude<CSSProperties['width'], undefined>
-  onMouseDownPoint: (mouse_start_x: number, point: number) => void
-}) {
-  function isHighlight(floor_idx: number) {
-    return highlightList.indexOf(floor_idx) !== -1
-  }
-
-  return (
-    <>
-      <div className={s.FloorTextLayout}>
-        {floors.map(floor_idx => {
-          return (
-            <div
-              key={floor_idx}
-              className={`${s.FloorText} ${isHighlight(floor_idx) ? s.FloorTextHighlight : ''}`}
-              style={{ left: `calc(${floor_idx} * ${intervalWidth} - ( var(--text-width) / 2 ))`
-            }}>{floor_idx + 1}F</div>
-          )
-        })}
-      </div>
-      <div className={s.FloorPointLayout}>
-        {floors.map(floor_idx => {
-          return (
-            <div
-              key={floor_idx}
-              className={`${s.FloorPoint} ${isHighlight(floor_idx) ? s.FloorPointHighlight : ''}`}
-              style={{ left: `calc(${floor_idx} * ${intervalWidth} - ( var(--point-size) / 2 ))`}}
-              onMouseDown={e => {
-                e.preventDefault()
-                e.stopPropagation()
-                onMouseDownPoint(e.clientX, floor_idx)
-              }}
-            ></div>
-          )
-        })}
-      </div>
-    </>
-  )
-}
-
-type SelectDirect = 'LEFT' | 'RIGHT'
-function Selected({
-  dragStartPoint,
-  dragEndPoint,
-  intervalWidth,
-  backgroundColor,
-}: {
-  dragStartPoint: number
-  dragEndPoint: number
-  intervalWidth: number
-  backgroundColor?: CSSProperties['backgroundColor']
-}) {
-  const direct = useMemo<SelectDirect>(() => {
-    if (dragEndPoint > dragStartPoint) {
-      return 'RIGHT'
-    } else {
-      return 'LEFT'
-    }
-  }, [dragEndPoint, dragStartPoint])
-
-  const selectedWidth = useMemo(() => {
-    const len = Math.abs(dragEndPoint - dragStartPoint)
-    return len * intervalWidth
-  }, [dragEndPoint, dragStartPoint, intervalWidth])
-
-  const selectedLeft = useMemo(() => {
-    if (direct === 'LEFT') {
-      return dragEndPoint * intervalWidth
-    } else {
-      return dragStartPoint * intervalWidth
-    }
-  }, [direct, dragEndPoint, dragStartPoint, intervalWidth])
-
-  return (
-    <div className={s.SelectedContainer} style={{ backgroundColor }}>
-      <div
-        className={s.Selected}
-        style={{
-          width: `${selectedWidth}px`,
-          left: `${selectedLeft}px`
-        }}
-      ></div>
-    </div>
-  )
 }
 
 type SelectedFloors = number[]
